Fixed returns of assigning binary expressions causing syntax errors, moved tests
